+++ core/src/main/java/com/zextras/carbonio/files/config/FilesConfig.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.carbonio.files.config.interfaces;
+package com.zextras.carbonio.files.config;
 
 import com.zextras.carbonio.preview.PreviewClient;
 import com.zextras.carbonio.usermanagement.UserManagementClient;
