+++ core/src/test/java/com/zextras/carbonio/files/clients/DocsConnectorHttpClientTest.java
@@ -4,7 +4,7 @@
 
 package com.zextras.carbonio.files.clients;
 
-import com.zextras.carbonio.files.config.interfaces.FilesConfig;
+import com.zextras.carbonio.files.config.FilesConfig;
 import java.io.IOException;
 import org.apache.http.ProtocolVersion;
 import org.apache.http.client.methods.CloseableHttpResponse;
