+++ core/src/test/java/com/zextras/carbonio/files/dal/FilesPostgreSQLContainer.java
@@ -4,7 +4,7 @@
 
 package com.zextras.carbonio.files.dal;
 
-import com.zextras.carbonio.files.config.interfaces.FilesConfig;
+import com.zextras.carbonio.files.config.FilesConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.containers.PostgreSQLContainer;
